+++ src/UpdateVisuals.ts
@@ -744,3 +744,19 @@ export const visualUpdateShop = () => {
     DOMCacheGetOrSet('singularityQuarksLevel').textContent = `${player.goldenQuarks < 1000 ? 'Owned: ' : ''}${format(player.goldenQuarks)}`
     DOMCacheGetOrSet('singularityQuarksButton').textContent = `Buy! ${format(getGoldenQuarkCost().cost)} Quarks Each`
 }
+
+export const shopOthersDescriptions = (input: ShopUpgradeNames) => {
+    const rofl = DOMCacheGetOrSet('quarkdescription')!;
+    const lol = DOMCacheGetOrSet('quarkeffect')!;
+    DOMCacheGetOrSet('quarkRefundable')!.textContent = 'This item CANNOT be refunded! Take caution.';
+
+    const goldenQuarkCost = getGoldenQuarkCost();
+    const maxBuy = Math.floor(+player.worlds / goldenQuarkCost.cost);
+
+    switch (input) {
+        case 'singularityQuarks':
+            rofl.textContent = 'You can buy Golden Quarks at Quarks. If you don\'t like Singularity, trade!'
+            lol.textContent = `You can buy Golden Quarks here for ${format(goldenQuarkCost.cost, 0, true)} Quarks (Discounted by ${format(goldenQuarkCost.costReduction, 0, true)})! You can buy up to ${format(maxBuy, 0, true)}.`;
+            break;
+    }
+}
